+++ CHANGES.md
@@ -1,3 +1,8 @@
+# Unreleased
+- [DEPRECATED] `DatastoreManager` constructors. Use `DatastoreManager.getInstance` factory methods
+  instead to guarantee only a single `DatastoreManager` instance is created for a given storage
+  directory path in the scope of the `DatastoreManager` class.
+
 # 1.0.0 (2016-05-03)
 - [NOTE] This library follows the
   [Semantic Versioning 2.0.0 specification](http://semver.org). To